Move specs.py to its own module
This allows to have a SpecRule library
