+++ src/tqec/compile/specs/__init__.py
@@ -1,0 +1,4 @@
+from .base import CubeSpec, SpecRule
+
+# TODO: Only temporary, removing soon
+from .library.css import DEFAULT_SPEC_RULES
